+++ fragments/be_password/reset.php
@@ -1,30 +1,22 @@
 <?php
 
-// check for new login page that comes with fragment for background image
-$isNewLoginPage = file_exists(rex_path::core('fragments/core/login_background.php'));
+use FriendsOfRedaxo\BePassword\Services\RenderService;
 
-$email = rex_request('email', 'string');
-$content = $successMessage = $errorMessage = $buttons = '';
-$showForm = isset($showForm) ? (bool) $showForm : false;
-$token = isset($token) ? (string) $token : '';
+/** @var RenderService $this */
 
-$success = isset($success) ? (string) $success : '';
-$error = isset($error) ? (string) $error : '';
+$email = $this->getEmail();
+$content = $buttons = '';
+$showForm = $this->getShowForm();
+$token = $this->getToken();
 
+$success = $this->getSuccessMsg();
 if ('' < $success) {
-    $successMessage = '<div class="rex-js-login-message">' . rex_view::success($success) . '</div>';
+    $content .= '<div class="rex-js-login-message">' . rex_view::success($success) . '</div>';
 }
 
+$error = $this->getErrorMsg();
 if ('' < $error) {
-    $errorMessage = '<div class="rex-js-login-message">' . rex_view::error($error) . '</div>';
-}
-
-if (!$isNewLoginPage) {
-    echo $successMessage;
-    echo $errorMessage;
-} else {
-    $content .= $successMessage;
-    $content .= $errorMessage;
+    $content .= '<div class="rex-js-login-message">' . rex_view::error($error) . '</div>';
 }
 
 if ('' === $success || true === $showForm) {
@@ -68,7 +60,7 @@ if ('' === $success || true === $showForm) {
     $buttons = $fragment->parse('core/form/submit.php');
 }
 
-if (!('' < $success && !$isNewLoginPage) || true === $showForm) {
+if (true === $showForm) {
     $fragment = new rex_fragment();
     $fragment->setVar('title', rex_i18n::msg('be_password_select_new_password'), false);
     $fragment->setVar('body', $content, false);